Subtract scrollbar in PerpendicularContainingBlockLogicalHeight
This is an issue that happens in Flexbox and Grid Layout
and any other layout method that uses OverrideLogicalHeight.
In LayoutBox::PerpendicularContainingBlockLogicalHeight()
we were returning the OverrideContentLogicalHeight() plus
the scrollbar height.
That caused issue with orthogonal flows between flex/grid items
and their children.

The issue is fixed just following the suggestion in the TODO
and using OverrideContentLogicalHeight()
(which doesn't include the scrollbar height).

BUG=836827
TEST=external/wpt/css/css-flexbox/percentage-size-subitems-001.html
TEST=external/wpt/css/css-grid/grid-items/percentage-size-subitems-001.html

Change-Id: I0c7586449067aab78430defc831b2774136e2f0b